+++ minidns-core/src/main/java/de/measite/minidns/source/DNSDataSource.java
@@ -8,13 +8,13 @@
  * upon the condition that you accept all of the terms of either
  * the Apache License 2.0, the LGPL 2.1+ or the WTFPL.
  */
-package de.measite.minidns.world;
+package de.measite.minidns.source;
 
 import de.measite.minidns.DNSMessage;
 
 import java.net.InetAddress;
 
-public abstract class DNSWorld {
+public abstract class DNSDataSource {
     public abstract DNSMessage query(DNSMessage message, InetAddress address, int port);
 
     private int udpPayloadSize = 512;
